ecom-chatbot/requirements.txt
accelerate==0.24.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bitsandbytes==0.41.1
bleach==6.1.0
blinker==1.7.0
blobfile==2.0.1
bs4==0.0.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
curl-cffi==0.5.10
cycler==0.11.0
dataclasses-json==0.6.1
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
duckduckgo_search==4.1.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.0
fastjsonschema==2.19.1
filelock==3.9.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.10.0
google-auth==2.23.3
googleapis-common-protos==1.61.0
gpt4all==2.0.1
greenlet==3.0.1
grpcio==1.59.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.20.1
ipython==8.10.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
kubernetes==28.1.0
langchain==0.0.330
langchain-core==0.1.7
langchain-openai==0.0.2
langsmith==0.0.77
loguru==0.7.2
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.8
numpy==1.24.1
oauthlib==3.2.2
onnxruntime==1.16.1
openai==0.28.1
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.4.12
platformdirs==2.6.2
plotly==5.12.0
posthog==3.0.2
prometheus-client==0.19.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.4
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.17
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.15.0
pyparsing==3.0.9
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.9
safetensors==0.4.0
scikit-learn==1.2.0
scipy==1.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.22
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.1.0
terminado==0.18.0
threadpoolctl==3.1.0
tiktoken==0.1.2
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
torch==2.1.2
tornado==6.3.3
tqdm==4.64.1
traitlets==5.8.1
transformers==4.30.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.4.0
uri-template==1.3.0
urllib3==1.26.17
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wikipedia==1.4.0
wrapt==1.15.0
yarg==0.1.9
yarl==1.8.2
zipp==3.17.0
